requirements.txt
langgraph==0.0.20
langchain==0.2.0
langchain-community==0.2.0
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.29.0
python-dotenv==1.0.1
pydantic==2.5.3
pytest==7.4.4
duckduckgo-search==5.0.0
pypdf==4.0.0
sqlalchemy==2.0.25